backend/requirements.txt
# OhmVision Platform - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet>=3.0.0

# Cache & Queue
redis==5.0.1
celery==5.3.6

# AI & Computer Vision
ultralytics>=8.2.0
opencv-python==4.9.0.80
numpy==1.26.3
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0  # CPU version (GPU version not available on Mac ARM)
pillow==10.2.0

# LLM / Agent IA
anthropic==0.18.0
openai==1.12.0
langchain==0.1.6

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
jinja2==3.1.3

# Email
aiosmtplib==3.0.1
emails==0.6

# Payments (Stripe)
stripe==8.0.0
email-validator>=2.0.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
aiosqlite==0.19.0

# WebSocket
websockets==12.0

# PDF Reports
reportlab==4.0.7

# Date utilities
python-dateutil==2.8.2

# System monitoring
psutil==5.9.8
